+++ sdk/go/aws/neptune/subnetGroup.go
@@ -22,8 +22,8 @@ func NewSubnetGroup(ctx *pulumi.Context,
 		return nil, errors.New("missing required argument 'SubnetIds'")
 	}
 	inputs := make(map[string]interface{})
+	inputs["description"] = "Managed by Pulumi"
 	if args == nil {
-		inputs["description"] = nil
 		inputs["name"] = nil
 		inputs["namePrefix"] = nil
 		inputs["subnetIds"] = nil