+++ sdk/python/pulumi_aws/appsync/api_key.py
@@ -58,6 +58,8 @@ class ApiKey(pulumi.CustomResource):
             if api_id is None:
                 raise TypeError("Missing required property 'api_id'")
             __props__['api_id'] = api_id
+            if description is None:
+                description = 'Managed by Pulumi'
             __props__['description'] = description
             __props__['expires'] = expires
             __props__['key'] = None